Extract target skill installation
Move target-root install, cleanup, deprecated-skill removal, and managed directory pruning into SkillTargetInstaller. This leaves ToolSkillSynchronizer focused on orchestration and result aggregation.
