+++ Packages/src/Editor/Infrastructure/SkillSetup/SkillTargetInstaller.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: baa8d9096e6264b589a54056ec16544d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
